feat(github): Improve check run output length detection
